Refactor/reorganize + omit final slash for dirname
I'm toying with the right approach for exporting the virtual module.
Augmentations/ambient decls are out for JSR support (I think), and
I don't want processed code to be importing from a "plugin" module,
so I've inverted the relationship between the project and the plugin
so that the plugin is a submodule and the virtual module lives at the
root.

Also ejected project.json for now until I figure out the right setup
for JSR publishing and associated interactions with client projects.